module AuthService -> class AuthService
In the controller (specifically, in the file
controller/lib/openshift-origin-controller/lib/openshift/auth_service.rb),
AuthService is declared as a class.  Prior to this patch, the plug-uns
under plugins/auth all attempt to re-open this class as a module, which
fails with "AuthService is not a module."

This patch changes these plug-ins to re-open the class as a class so
that the code works.