ActiveMQ routing plug-in: fix initialisation
Open the connection to ActiveMQ in the initialize method of
ActiveMQPlugin instead of using a conditional assignment in send_msg.

To open this connection, we need values from Rails.application.config,
which is nil at the time that openshift-origin-routing-activemq.rb is
loaded.  Therefore, this commit also moves the instantiation of
ActiveMQPlugin from openshift-origin-routing-activemq.rb to an
after_initialize callback for ActiveMQRoutingEngine.
Rails.application.config is defined by the time the callback is invoked.